fix(AutoComplete): fix keyboard interaction options not working on AutoComplete
