+++ src/Picker/index.ts
@@ -7,7 +7,6 @@ import {
 
 import { PickerToggleProps } from './PickerToggle';
 import { PickerHandle, PickerComponent } from './types';
-
 export { default as TreeView } from './TreeView';
 export { default as DropdownMenu } from './DropdownMenu';
 export { default as DropdownMenuCheckItem } from './DropdownMenuCheckItem';
@@ -20,8 +19,6 @@ export { default as PickerIndicator } from './PickerIndicator';
 export { default as PickerToggleTrigger } from './PickerToggleTrigger';
 export { default as SearchBar } from './SearchBar';
 export { default as SelectedElement } from './SelectedElement';
-export { default as usePickerRef } from './usePickerRef';
-export { default as useCombobox } from './useCombobox';
 export { pickTriggerPropKeys, omitTriggerPropKeys };
 
 export type {
@@ -31,5 +28,7 @@ export type {
   PickerComponent,
   PickerToggleProps
 };
+
+export * from './hooks';
 export * from './utils';
 export * from './propTypes';
