+++ src/Picker/utils.ts
@@ -1,23 +1,7 @@
-import React, { useState, useCallback, useMemo } from 'react';
-import kebabCase from 'lodash/kebabCase';
+import React from 'react';
 import trim from 'lodash/trim';
-import isFunction from 'lodash/isFunction';
-import isUndefined from 'lodash/isUndefined';
-import omit from 'lodash/omit';
-import find from 'lodash/find';
-import { OverlayTriggerHandle } from './PickerToggleTrigger';
 import { findNodeOfTree } from '../utils/treeUtils';
-import {
-  KEY_VALUES,
-  useClassNames,
-  shallowEqual,
-  reactToString,
-  placementPolyfill,
-  useEventCallback
-} from '../utils';
-import { TypeAttributes } from '../@types/common';
-import type { ListHandle } from '../Windowing';
-import { getHeight } from 'dom-lib';
+import { KEY_VALUES, reactToString } from '../utils';
 
 export interface NodeKeys {
   valueKey: string;
@@ -56,65 +40,6 @@ export function shouldDisplay(label: React.ReactNode, searchKeyword: string) {
     return nodes.join('').toLocaleLowerCase().indexOf(keyword) >= 0;
   }
   return false;
-}
-
-export interface PickerClassNameProps {
-  name?: string;
-  classPrefix: string;
-  className?: string;
-  placement?: TypeAttributes.Placement;
-  appearance?: 'default' | 'subtle';
-  cleanable?: boolean;
-  block?: boolean;
-  disabled?: boolean;
-  countable?: boolean;
-  readOnly?: boolean;
-  plaintext?: boolean;
-  hasValue?: boolean;
-  classes?: any;
-}
-
-/**
- * The className of the assembled Toggle is on the Picker.
- */
-export function usePickerClassName(props: PickerClassNameProps): [string, string[]] {
-  const {
-    name,
-    classPrefix,
-    className,
-    placement,
-    appearance,
-    cleanable,
-    block,
-    disabled,
-    countable,
-    readOnly,
-    plaintext,
-    hasValue,
-    ...rest
-  } = props;
-
-  const { withClassPrefix, merge } = useClassNames(classPrefix);
-  const classes = merge(
-    className,
-    withClassPrefix(name, appearance, 'toggle-wrapper', {
-      [`placement-${kebabCase(placementPolyfill(placement))}`]: placement,
-      'read-only': readOnly,
-      'has-value': hasValue,
-      cleanable,
-      block,
-      disabled,
-      countable,
-      plaintext
-    })
-  );
-
-  // Those props that're used for composing the className
-  const usedClassNamePropKeys = Object.keys(
-    omit(props, [...Object.keys(rest || {}), 'disabled', 'readOnly', 'plaintext', 'name'])
-  );
-
-  return [classes, usedClassNamePropKeys];
 }
 
 export interface EventsProps {
@@ -169,465 +94,4 @@ export function onMenuKeyDown(event: React.KeyboardEvent, events: EventsProps) {
       break;
     default:
   }
-}
-
-export interface FocusItemValueProps<T = unknown> {
-  target: HTMLElement | null | (() => HTMLElement | null);
-  data?: T[];
-  valueKey?: string;
-  focusableQueryKey?: string;
-  defaultLayer?: number;
-  rtl?: boolean;
-  callback?: (value: any, event: React.KeyboardEvent) => void;
-  getParent?: (node: T) => T | undefined;
-}
-
-/**
- * Checks if the element has a vertical scrollbar.
- */
-function hasVerticalScroll(element: HTMLElement) {
-  const { scrollHeight, clientHeight } = element;
-  return scrollHeight > clientHeight;
-}
-
-/**
- * Checks if the element is within the visible area of the container
- */
-function isVisible(element: HTMLElement, container: HTMLElement, direction: 'top' | 'bottom') {
-  if (!hasVerticalScroll(container)) {
-    return true;
-  }
-
-  const { top, bottom, height } = element.getBoundingClientRect();
-  const { top: containerTop, bottom: containerBottom } = container.getBoundingClientRect();
-
-  if (direction === 'top') {
-    return top + height > containerTop;
-  }
-  return bottom - height < containerBottom;
-}
-
-function scrollTo(container: HTMLElement, direction: 'top' | 'bottom', step: number) {
-  const { scrollTop } = container;
-  container.scrollTop = direction === 'top' ? scrollTop - step : scrollTop + step;
-}
-
-/**
- * A hook that manages the focus state of the option.
- * @param defaultFocusItemValue
- * @param props
- */
-export const useFocusItemValue = <T, D>(
-  defaultFocusItemValue: T | null | undefined,
-  props: FocusItemValueProps<D>
-) => {
-  const {
-    valueKey = 'value',
-    focusableQueryKey = '[data-key][aria-disabled="false"]',
-    defaultLayer = 0,
-    data,
-    target,
-    rtl,
-    callback,
-    // TODO-Doma This legacy behavior of using `.parent` property should be deprecated
-    //           Always explicitly pass `getParent` when there's need to traverse upwards
-    getParent = item => (item as any)?.parent
-  } = props;
-  const [focusItemValue, setFocusItemValue] = useState<T | null | undefined>(defaultFocusItemValue);
-  const [layer, setLayer] = useState(defaultLayer);
-  const [keys, setKeys] = useState<any[]>([]);
-
-  const focusCallback = useCallback(
-    (value: any, event: React.KeyboardEvent) => {
-      const menu = isFunction(target) ? target() : target;
-      const focusElement = menu?.querySelector(`[data-key="${value}"]`) as HTMLElement;
-
-      focusElement?.focus();
-      callback?.(value, event);
-    },
-    [callback, target]
-  );
-
-  const getScrollContainer = useCallback(() => {
-    const menu = isFunction(target) ? target() : target;
-
-    // For Cascader and MutiCascader
-    const subMenu = menu?.querySelector(`[data-layer="${layer}"]`);
-
-    if (subMenu) {
-      return subMenu;
-    }
-
-    // For SelectPicker、CheckPicker、Autocomplete、InputPicker、TagPicker
-    return menu?.querySelector('[role="listbox"]');
-  }, [layer, target]);
-
-  /**
-   * Get the elements visible in all options.
-   */
-  const getFocusableMenuItems = useCallback(() => {
-    if (!target) {
-      return [];
-    }
-
-    let currentKeys = keys;
-
-    if (layer < 1) {
-      const popup = isFunction(target) ? target() : target;
-
-      const rootMenu = popup?.querySelector<HTMLElement>('[data-layer="0"]');
-
-      if (rootMenu) {
-        currentKeys = Array.from(
-          rootMenu.querySelectorAll<HTMLElement>(focusableQueryKey) ?? []
-        ).map(item => item.dataset?.key);
-      } else {
-        currentKeys = Array.from(popup?.querySelectorAll<HTMLElement>(focusableQueryKey) ?? []).map(
-          item => item.dataset?.key
-        );
-      }
-    }
-
-    // 1. It is necessary to traverse the `keys` instead of `data` here to preserve the order of the array.
-    // 2. The values ​​in `keys` are all string, so the corresponding value of `data` should also be converted to string
-    return currentKeys.map(key => find(data, i => `${i[valueKey]}` === key));
-  }, [data, focusableQueryKey, keys, target, valueKey, layer]);
-
-  /**
-   * Get the index of the focus item.
-   */
-  const findFocusItemIndex = useCallback(
-    callback => {
-      const items = getFocusableMenuItems();
-
-      for (let i = 0; i < items.length; i += 1) {
-        if (shallowEqual(focusItemValue, items[i]?.[valueKey])) {
-          callback(items, i);
-          return;
-        }
-      }
-      callback(items, -1);
-    },
-    [focusItemValue, getFocusableMenuItems, valueKey]
-  );
-
-  const scrollListItem = useCallback(
-    (direction: 'top' | 'bottom', itemValue: string, willOverflow: boolean) => {
-      const container = getScrollContainer() as HTMLElement;
-      const item = container?.querySelector<HTMLElement>(`[data-key="${itemValue}"]`);
-
-      if (willOverflow && container) {
-        const { scrollHeight, clientHeight } = container;
-        container.scrollTop = direction === 'top' ? scrollHeight - clientHeight : 0;
-        return;
-      }
-
-      if (item && container) {
-        if (!isVisible(item, container, direction)) {
-          const height = getHeight(item);
-
-          scrollTo(container, direction, height);
-        }
-      }
-    },
-    [getScrollContainer]
-  );
-
-  const focusNextMenuItem = useCallback(
-    (event: React.KeyboardEvent) => {
-      findFocusItemIndex((items, index) => {
-        const willOverflow = index + 2 > items.length;
-        const nextIndex = willOverflow ? 0 : index + 1;
-        const focusItem = items[nextIndex];
-
-        if (!isUndefined(focusItem)) {
-          setFocusItemValue(focusItem[valueKey]);
-          focusCallback(focusItem[valueKey], event);
-          scrollListItem('bottom', focusItem[valueKey], willOverflow);
-        }
-      });
-    },
-    [focusCallback, findFocusItemIndex, scrollListItem, valueKey]
-  );
-
-  const focusPrevMenuItem = useCallback(
-    (event: React.KeyboardEvent) => {
-      findFocusItemIndex((items, index) => {
-        const willOverflow = index === 0;
-        const nextIndex = willOverflow ? items.length - 1 : index - 1;
-        const focusItem = items[nextIndex];
-        if (!isUndefined(focusItem)) {
-          setFocusItemValue(focusItem[valueKey]);
-          focusCallback(focusItem[valueKey], event);
-          scrollListItem('top', focusItem[valueKey], willOverflow);
-        }
-      });
-    },
-    [focusCallback, findFocusItemIndex, scrollListItem, valueKey]
-  );
-
-  const getSubMenuKeys = useCallback(
-    (nextLayer: number) => {
-      const menu = isFunction(target) ? target() : target;
-      const subMenu = menu?.querySelector(`[data-layer="${nextLayer}"]`);
-
-      if (subMenu) {
-        return Array.from(subMenu.querySelectorAll<HTMLElement>(focusableQueryKey))?.map<
-          T | undefined
-        >(item => item.dataset?.key as any);
-      }
-
-      return null;
-    },
-    [focusableQueryKey, target]
-  );
-
-  const focusNextLevelMenu = useCallback(
-    (event: React.KeyboardEvent) => {
-      const nextLayer = layer + 1;
-      const nextKeys = getSubMenuKeys(nextLayer);
-
-      if (nextKeys) {
-        setKeys(nextKeys);
-        setLayer(nextLayer);
-        setFocusItemValue(nextKeys[0]);
-        focusCallback(nextKeys[0], event);
-      }
-    },
-    [focusCallback, getSubMenuKeys, layer]
-  );
-
-  const focusPrevLevelMenu = useCallback(
-    (event: React.KeyboardEvent) => {
-      const nextLayer = layer - 1;
-      const nextKeys = getSubMenuKeys(nextLayer);
-
-      if (nextKeys) {
-        setKeys(nextKeys);
-        setLayer(nextLayer);
-
-        const focusItem = findNodeOfTree(data, item => item[valueKey] === focusItemValue);
-        const parentItemValue = getParent(focusItem)?.[valueKey];
-
-        if (parentItemValue) {
-          setFocusItemValue(parentItemValue);
-          focusCallback(parentItemValue, event);
-        }
-      }
-    },
-    [focusCallback, data, focusItemValue, getParent, getSubMenuKeys, layer, valueKey]
-  );
-
-  const handleKeyDown = useCallback(
-    (event: any) => {
-      onMenuKeyDown(event, {
-        down: focusNextMenuItem,
-        up: focusPrevMenuItem,
-        [rtl ? 'left' : 'right']: focusNextLevelMenu,
-        [rtl ? 'right' : 'left']: focusPrevLevelMenu
-      });
-    },
-    [focusNextLevelMenu, focusNextMenuItem, focusPrevLevelMenu, focusPrevMenuItem, rtl]
-  );
-
-  return {
-    focusItemValue,
-    setFocusItemValue,
-    layer,
-    setLayer,
-    keys,
-    setKeys,
-    onKeyDown: handleKeyDown
-  };
-};
-
-export interface ToggleKeyDownEventProps {
-  toggle?: boolean;
-  trigger: React.RefObject<any>;
-  target: React.RefObject<any>;
-  overlay?: React.RefObject<any>;
-  searchInput?: React.RefObject<any>;
-  active?: boolean;
-  readOnly?: boolean;
-  disabled?: boolean;
-  onExit?: (event) => void;
-  onKeyDown?: (event) => void;
-  onOpen?: () => void;
-  onClose?: () => void;
-  onMenuKeyDown?: (event) => void;
-  onMenuPressEnter?: (event) => void;
-  onMenuPressBackspace?: (event) => void;
-  [key: string]: any;
-}
-
-/**
- * A hook to control the toggle keyboard operation
- * @param props
- */
-export const useToggleKeyDownEvent = (props: ToggleKeyDownEventProps) => {
-  const {
-    toggle = true,
-    trigger,
-    target,
-    overlay,
-    searchInput,
-    active,
-    readOnly,
-    disabled,
-    loading,
-    onExit,
-    onOpen,
-    onClose,
-    onKeyDown,
-    onMenuKeyDown,
-    onMenuPressEnter,
-    onMenuPressBackspace
-  } = props;
-
-  const handleClose = useEventCallback(() => {
-    trigger.current?.close?.();
-
-    // The focus is on the trigger button after closing
-    trigger.current?.focus?.();
-    onClose?.();
-  });
-
-  const handleOpen = useEventCallback(() => {
-    trigger.current?.open?.();
-    onOpen?.();
-  });
-
-  const handleToggleDropdown = useEventCallback(() => {
-    if (active) {
-      handleClose();
-      return;
-    }
-    handleOpen();
-  });
-
-  const onToggle = useEventCallback((event: React.KeyboardEvent) => {
-    // Keyboard events should not be processed when readOnly and disabled are set.
-    if (readOnly || disabled || loading) {
-      return;
-    }
-
-    if (event.target === target?.current) {
-      // enter
-      if (toggle && event.key === KEY_VALUES.ENTER) {
-        handleToggleDropdown();
-      }
-
-      // delete
-      if (event.key === KEY_VALUES.BACKSPACE) {
-        onExit?.(event);
-      }
-    }
-
-    if (overlay?.current) {
-      // The keyboard operation callback on the menu.
-      onMenuKeyDown?.(event);
-
-      if (event.key === KEY_VALUES.ENTER) {
-        onMenuPressEnter?.(event);
-      }
-
-      /**
-       * There is no callback when typing the Backspace key in the search box.
-       * The default is to remove search keywords
-       */
-      if (event.key === KEY_VALUES.BACKSPACE && event.target !== searchInput?.current) {
-        onMenuPressBackspace?.(event);
-      }
-
-      // The search box gets focus when typing characters and numbers.
-      if (event.key.length === 1 && /\w/.test(event.key)) {
-        // Exclude Input
-        // eg: <SelectPicker renderExtraFooter={() => <Input />} />
-        if ((event.target as HTMLInputElement)?.tagName !== 'INPUT') {
-          searchInput?.current?.focus();
-        }
-      }
-    }
-
-    if (event.key === KEY_VALUES.ESC || event.key === KEY_VALUES.TAB) {
-      handleClose();
-    }
-
-    // Native event callback
-    onKeyDown?.(event);
-  });
-
-  return onToggle;
-};
-
-interface SearchOptions<T> {
-  labelKey: string;
-  searchBy?: (keyword: string, label: any, item: T) => boolean;
-  callback?: (keyword: string, data: T[], event: React.SyntheticEvent) => void;
-}
-
-type UseSearchResult<T> = {
-  searchKeyword: string;
-  filteredData: T[];
-  checkShouldDisplay: (item: T, keyword?: string) => boolean;
-  handleSearch: (searchKeyword: string, event: React.SyntheticEvent) => void;
-  resetSearch: () => void;
-};
-
-/**
- * A hook that handles search filter options
- */
-export function useSearch<T>(data: readonly T[], props: SearchOptions<T>): UseSearchResult<T> {
-  const { labelKey, searchBy, callback } = props;
-
-  // Use search keywords to filter options.
-  const [searchKeyword, setSearchKeyword] = useState('');
-
-  const resetSearch = useCallback(() => {
-    setSearchKeyword('');
-  }, []);
-
-  /**
-   * Index of keyword  in `label`
-   * @param {node} label
-   */
-  const checkShouldDisplay = useCallback(
-    (item: T, keyword?: string) => {
-      const checkValue = typeof item === 'object' ? item?.[labelKey] : String(item);
-      const _keyword = isUndefined(keyword) ? searchKeyword : keyword;
-
-      if (typeof searchBy === 'function') {
-        return searchBy(_keyword, checkValue, item);
-      }
-      return shouldDisplay(checkValue, _keyword);
-    },
-    [labelKey, searchBy, searchKeyword]
-  );
-
-  const filteredData = useMemo(() => {
-    return data.filter(item => checkShouldDisplay(item, searchKeyword));
-  }, [checkShouldDisplay, data, searchKeyword]);
-
-  const handleSearch = (searchKeyword: string, event: React.SyntheticEvent) => {
-    const filteredData = data.filter(item => checkShouldDisplay(item, searchKeyword));
-    setSearchKeyword(searchKeyword);
-    callback?.(searchKeyword, filteredData, event);
-  };
-
-  return {
-    searchKeyword,
-    filteredData,
-    checkShouldDisplay,
-    handleSearch,
-    resetSearch
-  };
-}
-
-export interface PickerDependentParameters {
-  triggerRef?: React.RefObject<OverlayTriggerHandle>;
-  rootRef?: React.RefObject<HTMLElement>;
-  overlayRef?: React.RefObject<HTMLElement>;
-  targetRef?: React.RefObject<HTMLElement>;
-  listRef?: React.RefObject<ListHandle>;
-  inline?: boolean;
 }
